libproxyproto_connect: abort on socket errors
If getting or setting flags on the socket fails:

* abort the connection
* return an error

Otherwise the behaviour of the connection is undefined, i.e., the
socket may still be non-blocking, causing connect(2) to return
EINPROGRESS.

The errno value is from fnctl(2): it is possible this value differs from
connect(2) and the result will not be expected by the caller. Reviewing
the man pages, it seems F_GETFL/F_SETFL should only fail if the fd is
invalid (EBADF).