Default to allowing connection without header
Follow the behaviour of haproxy and allow connections with and without
the proxy protocol header by default.

Remove the LIBPROXYPROTO_PROTOCOL_HEADER_IS_OPTIONAL environment
variable (which is now the default) and replace it with
LIBPROXYPROTO_MUST_USE_PROTOCOL_HEADER.  When set, connections without
the proxy protocol header are dropped.